engineb: MODULE M-DELAY -- exact (null 30/30 EXACTLY 0), 207 cyc/sample on S3, and the ring does not fit
engine_b/eb_delay.c + .h implement the DELAY TYPE 0 stage from the measured
specification in docs/engineb/FX_DELAY.md, with engine B's own state: one
struct, one COMPILE-TIME ring, no 12 MB context.

MEASURED
  null_b.py --module delay            30/30 EXACTLY 0, all 17 idle-prefix
                                      scenarios included
  engine_b/test_delay_ref.c           14/14 configurations EXACTLY 0 against
                                      the literal transcription, covering the
                                      branches no factory patch reaches
                                      (HIGH CUT bypass, the OFF gate, MUTE, a
                                      DELAY TIME that MOVES, the routing
                                      change, both feedback extremes)
  double -> float in the tap index    0 mismatches over 2,464,696,322 float32
                                      bit patterns; 0 softfloat calls result

NON-VACUITY, measured by accident and then on purpose
  The first version read the delay-time smoother's 4-cell pipeline as a 2-cell
  smoother and failed 15 of 30 scenarios at -33.9 dB. Planted afterwards:
  tap index off by one 15/30 at -38.4 dB; feedback taken before the loop
  damping 6/30 at -13.6 dB; mute fade not one sample late 4/30 at -81.7 dB.
  CALIBRATION: 1 ULP on wet = -146.7 dB (passes, 47 dB of headroom); the
  "obvious" interpolation regrouping t1+f*(t2-t1) = -152.5 dB, a REAL error the
  gate would not catch, so it is not used; dropping the two measured-1.0
  multiplies is EXACTLY 0, so that one is free.

COST [MODELED, cost.py measure --calls eb_delay_process=1]
  ESP32-S3, 48 kHz build (EB_DELAY_BIQUAD=0)  207 cyc/sample, band 132..529
                                              = 5.9% of the 3,500 budget
  ESP32-S3, 44.1 kHz build (biquad live)      235, band 149..605
  Cortex-M7 190 / host 57.  0 libm, 0 softfloat on all three.

MEMORY -- the finding that matters
  sizeof(eb_delay_state) = 524,400 B at EB_DELAY_LEN 65536 (manual 800 ms);
  1,048,688 B at 131072 (tempo sync 1/1 at 128 BPM). The 200 KB internal budget
  is exceeded 2.6x by the default ring alone. 6 ring accesses/sample: in
  internal SRAM the module is 207 cyc/sample, in PSRAM ~387/567/927
  [MODELED-UNVALIDATED, 30/60/120 cyc per access] -- 16% of the WHOLE engine
  budget at the nominal. The S3 PSRAM latency has never been measured by this
  project and is now the largest unknown in the delay.

SCOPE: DELAY TYPE 0 only. TYPES 1..5 are not written. docs/engineb/data/
eb_delay_cheap.c is superseded -- it is off by one sample in the feedback tap.

Full detail: docs/engineb/M-DELAY_RESULT.md

+++ engine_b/eb_delay.c
@@ -89,9 +89,18 @@ void eb_delay_process(const eb_delay_cfg *c, eb_delay_state *s,
     s->t_smooth = sm;
 
     /* ---- tap position: integer part NEGATED, fraction from the positive */
-    di2  = (int32_t)(float)(sm * -16384.0);        /* <= 0 */
-    P    = (float)(sm * 16384.0);
-    frac = (float)((double)P - (double)(int)P);
+    /* The sealed engine does these three in DOUBLE. Engine B does them in
+     * float, and that is not an assumption: the two forms were compared over
+     * EVERY float32 bit pattern with |sm| <= 1e6 -- 2,464,696,322 values --
+     * with 0 mismatches in the int cast and 0 in the fraction (the scaling is
+     * by a power of two, so it is exact, and P - trunc(P) is a difference of
+     * neighbouring floats). Values outside that domain set `overrun` below and
+     * are refused, so the untested tail is not reachable. This matters on the
+     * ESP32-S3, whose FPU is single-precision only: one double here would be a
+     * softfloat call per sample. */
+    di2  = (int32_t)(sm * -16384.0f);              /* <= 0 */
+    P    = sm * 16384.0f;
+    frac = P - (float)(int)P;
     if (-di2 + 2 >= EB_DELAY_LEN) s->overrun = 1;  /* never wrap silently */
 
     /* ---- read, loop damping filter, output, write ---------------------- */
